+++ Training/source/Rendering/IndexBuffer.h
@@ -5,14 +5,14 @@
 class index_buffer
 {
 public:
-    index_buffer(const void* data, unsigned int count);
+    index_buffer(const void* data, size_t count);
     ~index_buffer();
 
     void bind() const;
     void unbind() const;
-    inline unsigned int get_count() const { return m_count; }
+    inline size_t get_count() const { return m_count; }
 
 private:
     unsigned int m_renderer_id;
-    unsigned int m_count;
+    size_t m_count;
 };
